dom/bindings/test/test_bug742191.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=742191
-->
<head>
  <meta charset="utf-8">
  <title>Test for invalid argument object</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=742191">Mozilla Bug 742191</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 742191 **/
function doTest() {
  var gotTypeError = false;
  var ctx = document.createElement("canvas").getContext("2d");
  try {
    ctx.drawImage({}, 0, 0);
  } catch(e) {
    if (e instanceof TypeError) {
      gotTypeError = true;
    }
  }

  ok(gotTypeError, "passing an invalid argument should cause a type error!");
}
doTest();
</script>
</pre>
</body>
</html>
